requirements.txt
Django>=4.2,<5.3
psycopg2-binary>=2.9
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
mozilla-django-oidc>=4.0
africastalking>=1.2
gunicorn>=21.2
coverage>=7.5
django-mptt>=0.16
pytest>=8.2
pytest-django>=4.8
django-allauth>=65.11.0
